Performance Improvement
Worked on the Performance improvement on the image provider class.
Changed queue from serial to concurrent, for concurrent execution of image download.
Added new method in the Network layer to download the image data from given url.
The existing logic to download image as file required additional step to create image from the location url which was hampering the performance of the app.